Working in refactoring
