+++ reflexy/constants.py
@@ -1,6 +1,6 @@
 # GAME TIME
 GAME_SPEED = 10
-CLOCK_TICK_GAME_SPEED = 60
+CLOCK_TICK_GAME_SPEED = 30
 CLOCK_TICK_REFERENCE = 30
 
 # SCREEN
